+++ src/main/java/chess/ChessApplication.java
@@ -1,5 +1,7 @@
 package chess;
 
+import chess.controller.ChessController;
+
 public class ChessApplication {
     public static void main(String[] args) {
         new ChessController().run();
